+++ bgs_server/test/integration/test_tools/serverpod_test_tools.dart
@@ -37,13 +37,15 @@ import 'package:bgs_server/src/generated/organizations/models/organization_membe
 import 'dart:typed_data' as _i16;
 import 'package:bgs_server/src/generated/scheduling/models/scheduled_match.dart'
     as _i17;
-import 'package:bgs_server/src/generated/standings/models/standing.dart'
+import 'package:bgs_server/src/generated/search/models/search_results.dart'
     as _i18;
-import 'package:bgs_server/src/generated/teams/models/team.dart' as _i19;
+import 'package:bgs_server/src/generated/standings/models/standing.dart'
+    as _i19;
+import 'package:bgs_server/src/generated/teams/models/team.dart' as _i20;
 import 'package:bgs_server/src/generated/teams/models/team_membership.dart'
-    as _i20;
-import 'package:bgs_server/src/generated/teams/models/team_member_role.dart'
     as _i21;
+import 'package:bgs_server/src/generated/teams/models/team_member_role.dart'
+    as _i22;
 import 'package:bgs_server/src/generated/protocol.dart';
 import 'package:bgs_server/src/generated/endpoints.dart';
 export 'package:serverpod_test/serverpod_test_public_exports.dart';
@@ -176,6 +178,8 @@ class TestEndpoints {
 
   late final _ScheduledMatchEndpoint scheduledMatch;
 
+  late final _SearchEndpoint search;
+
   late final _StandingEndpoint standing;
 
   late final _TeamEndpoint team;
@@ -221,6 +225,10 @@ class _InternalTestEndpoints extends TestEndpoints
       serializationManager,
     );
     scheduledMatch = _ScheduledMatchEndpoint(
+      endpoints,
+      serializationManager,
+    );
+    search = _SearchEndpoint(
       endpoints,
       serializationManager,
     );
@@ -1775,6 +1783,52 @@ class _ScheduledMatchEndpoint {
   }
 }
 
+class _SearchEndpoint {
+  _SearchEndpoint(
+    this._endpointDispatch,
+    this._serializationManager,
+  );
+
+  final _i2.EndpointDispatch _endpointDispatch;
+
+  final _i2.SerializationManager _serializationManager;
+
+  _i3.Future<_i18.SearchResults> search(
+    _i1.TestSessionBuilder sessionBuilder, {
+    String? query,
+    _i9.Sport? sport,
+  }) async {
+    return _i1.callAwaitableFunctionAndHandleExceptions(() async {
+      var _localUniqueSession =
+          (sessionBuilder as _i1.InternalTestSessionBuilder).internalBuild(
+            endpoint: 'search',
+            method: 'search',
+          );
+      try {
+        var _localCallContext = await _endpointDispatch.getMethodCallContext(
+          createSessionCallback: (_) => _localUniqueSession,
+          endpointPath: 'search',
+          methodName: 'search',
+          parameters: _i1.testObjectToJson({
+            'query': query,
+            'sport': sport,
+          }),
+          serializationManager: _serializationManager,
+        );
+        var _localReturnValue =
+            await (_localCallContext.method.call(
+                  _localUniqueSession,
+                  _localCallContext.arguments,
+                )
+                as _i3.Future<_i18.SearchResults>);
+        return _localReturnValue;
+      } finally {
+        await _localUniqueSession.close();
+      }
+    });
+  }
+}
+
 class _StandingEndpoint {
   _StandingEndpoint(
     this._endpointDispatch,
@@ -1785,7 +1839,7 @@ class _StandingEndpoint {
 
   final _i2.SerializationManager _serializationManager;
 
-  _i3.Future<List<_i18.Standing>> listByLeague(
+  _i3.Future<List<_i19.Standing>> listByLeague(
     _i1.TestSessionBuilder sessionBuilder,
     _i2.UuidValue leagueId,
   ) async {
@@ -1808,7 +1862,7 @@ class _StandingEndpoint {
                   _localUniqueSession,
                   _localCallContext.arguments,
                 )
-                as _i3.Future<List<_i18.Standing>>);
+                as _i3.Future<List<_i19.Standing>>);
         return _localReturnValue;
       } finally {
         await _localUniqueSession.close();
@@ -1827,7 +1881,7 @@ class _TeamEndpoint {
 
   final _i2.SerializationManager _serializationManager;
 
-  _i3.Future<_i19.Team> create(
+  _i3.Future<_i20.Team> create(
     _i1.TestSessionBuilder sessionBuilder, {
     required _i2.UuidValue leagueId,
     required String name,
@@ -1854,7 +1908,7 @@ class _TeamEndpoint {
                   _localUniqueSession,
                   _localCallContext.arguments,
                 )
-                as _i3.Future<_i19.Team>);
+                as _i3.Future<_i20.Team>);
         return _localReturnValue;
       } finally {
         await _localUniqueSession.close();
@@ -1862,7 +1916,7 @@ class _TeamEndpoint {
     });
   }
 
-  _i3.Future<_i19.Team?> getById(
+  _i3.Future<_i20.Team?> getById(
     _i1.TestSessionBuilder sessionBuilder,
     _i2.UuidValue teamId,
   ) async {
@@ -1885,7 +1939,7 @@ class _TeamEndpoint {
                   _localUniqueSession,
                   _localCallContext.arguments,
                 )
-                as _i3.Future<_i19.Team?>);
+                as _i3.Future<_i20.Team?>);
         return _localReturnValue;
       } finally {
         await _localUniqueSession.close();
@@ -1893,7 +1947,7 @@ class _TeamEndpoint {
     });
   }
 
-  _i3.Future<List<_i19.Team>> listByLeague(
+  _i3.Future<List<_i20.Team>> listByLeague(
     _i1.TestSessionBuilder sessionBuilder,
     _i2.UuidValue leagueId,
   ) async {
@@ -1916,7 +1970,7 @@ class _TeamEndpoint {
                   _localUniqueSession,
                   _localCallContext.arguments,
                 )
-                as _i3.Future<List<_i19.Team>>);
+                as _i3.Future<List<_i20.Team>>);
         return _localReturnValue;
       } finally {
         await _localUniqueSession.close();
@@ -1924,7 +1978,7 @@ class _TeamEndpoint {
     });
   }
 
-  _i3.Future<List<_i20.TeamMembership>> listMembers(
+  _i3.Future<List<_i21.TeamMembership>> listMembers(
     _i1.TestSessionBuilder sessionBuilder,
     _i2.UuidValue teamId,
   ) async {
@@ -1947,7 +2001,7 @@ class _TeamEndpoint {
                   _localUniqueSession,
                   _localCallContext.arguments,
                 )
-                as _i3.Future<List<_i20.TeamMembership>>);
+                as _i3.Future<List<_i21.TeamMembership>>);
         return _localReturnValue;
       } finally {
         await _localUniqueSession.close();
@@ -1955,11 +2009,11 @@ class _TeamEndpoint {
     });
   }
 
-  _i3.Future<_i20.TeamMembership> invitePlayer(
+  _i3.Future<_i21.TeamMembership> invitePlayer(
     _i1.TestSessionBuilder sessionBuilder, {
     required _i2.UuidValue teamId,
     required String email,
-    _i21.TeamMemberRole? role,
+    _i22.TeamMemberRole? role,
   }) async {
     return _i1.callAwaitableFunctionAndHandleExceptions(() async {
       var _localUniqueSession =
@@ -1984,7 +2038,7 @@ class _TeamEndpoint {
                   _localUniqueSession,
                   _localCallContext.arguments,
                 )
-                as _i3.Future<_i20.TeamMembership>);
+                as _i3.Future<_i21.TeamMembership>);
         return _localReturnValue;
       } finally {
         await _localUniqueSession.close();
@@ -1992,7 +2046,7 @@ class _TeamEndpoint {
     });
   }
 
-  _i3.Future<_i20.TeamMembership> acceptInvite(
+  _i3.Future<_i21.TeamMembership> acceptInvite(
     _i1.TestSessionBuilder sessionBuilder,
     _i2.UuidValue membershipId,
   ) async {
@@ -2015,7 +2069,7 @@ class _TeamEndpoint {
                   _localUniqueSession,
                   _localCallContext.arguments,
                 )
-                as _i3.Future<_i20.TeamMembership>);
+                as _i3.Future<_i21.TeamMembership>);
         return _localReturnValue;
       } finally {
         await _localUniqueSession.close();
@@ -2023,7 +2077,7 @@ class _TeamEndpoint {
     });
   }
 
-  _i3.Future<_i20.TeamMembership> declineInvite(
+  _i3.Future<_i21.TeamMembership> declineInvite(
     _i1.TestSessionBuilder sessionBuilder,
     _i2.UuidValue membershipId,
   ) async {
@@ -2046,7 +2100,7 @@ class _TeamEndpoint {
                   _localUniqueSession,
                   _localCallContext.arguments,
                 )
-                as _i3.Future<_i20.TeamMembership>);
+                as _i3.Future<_i21.TeamMembership>);
         return _localReturnValue;
       } finally {
         await _localUniqueSession.close();
@@ -2054,7 +2108,7 @@ class _TeamEndpoint {
     });
   }
 
-  _i3.Future<List<_i20.TeamMembership>> listMine(
+  _i3.Future<List<_i21.TeamMembership>> listMine(
     _i1.TestSessionBuilder sessionBuilder,
   ) async {
     return _i1.callAwaitableFunctionAndHandleExceptions(() async {
@@ -2076,7 +2130,7 @@ class _TeamEndpoint {
                   _localUniqueSession,
                   _localCallContext.arguments,
                 )
-                as _i3.Future<List<_i20.TeamMembership>>);
+                as _i3.Future<List<_i21.TeamMembership>>);
         return _localReturnValue;
       } finally {
         await _localUniqueSession.close();